changes
git-svn-id: https://picviz.svn.sourceforge.net/svnroot/picviz/trunk@34 9f3d20f0-5e4f-0410-9e45-c526e915cc34

+++ src/libpicviz/include/types.h
@@ -18,7 +18,7 @@
 #ifndef _TYPES_H_
 #define _TYPES_H_
 
-typedef long long PcvHeight;
+typedef unsigned long PcvHeight;
 typedef unsigned int PcvWidth;
 typedef unsigned int PcvID;
 typedef char *       PcvString;
